fixup! Update airflow/models/baseoperator.py
