+++ tests/core/test_core.py
@@ -16,7 +16,6 @@
 # specific language governing permissions and limitations
 # under the License.
 
-import logging
 import os
 import signal
 from datetime import timedelta
@@ -371,46 +370,3 @@ class TestCore:
 
         assert context1['params'] == {'key_1': 'value_1', 'key_2': 'value_2_new', 'key_3': 'value_3'}
         assert context2['params'] == {'key_1': 'value_1', 'key_2': 'value_2_old'}
-
-
-def test_operator_retries_invalid(dag_maker):
-    with pytest.raises(AirflowException) as ctx:
-        with dag_maker():
-            BashOperator(
-                task_id='test_illegal_args',
-                bash_command='echo success',
-                retries='foo',
-            )
-        dag_maker.create_dagrun()
-    assert str(ctx.value) == "'retries' type must be int, not str"
-
-
-def test_operator_retries_coerce(caplog, dag_maker):
-    with caplog.at_level(logging.WARNING):
-        with dag_maker():
-            BashOperator(
-                task_id='test_illegal_args',
-                bash_command='echo success',
-                retries='1',
-            )
-        dag_maker.create_dagrun()
-    assert caplog.record_tuples == [
-        (
-            "airflow.operators.bash.BashOperator",
-            logging.WARNING,
-            "Implicitly converting 'retries' for <Task(BashOperator): test_illegal_args> from '1' to int",
-        ),
-    ]
-
-
-@pytest.mark.parametrize("retries", [None, 5])
-def test_operator_retries(caplog, dag_maker, retries):
-    with caplog.at_level(logging.WARNING):
-        with dag_maker(TEST_DAG_ID + str(retries)):
-            BashOperator(
-                task_id='test_illegal_args',
-                bash_command='echo success',
-                retries=retries,
-            )
-        dag_maker.create_dagrun()
-    assert caplog.records == []
